#52 - Use numbers as double insted of int

+++ src/parser/parser.cpp
@@ -431,7 +431,7 @@ Instruction* Parser::number(lexer::Token* tok, ON on) {
 
 Instruction* Parser::number_init(std::string val) {
 	Instruction* ins = new Instruction(Ins::LITERAL);
-	ins->value = new Value(Type::NUMBER, std::stoi(val));
+	ins->value = new Value(Type::NUMBER, std::stod(val));
 	return ins;
 }
